Fixed warnings

+++ Crystal_Engine_v2/src/crystal/imgui/ImGuiLayer.cpp
@@ -34,7 +34,7 @@ namespace Crystal
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& application = Application::Get();
-		io.DisplaySize = ImVec2(application.GetWindow().GetWidth(), application.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float) application.GetWindow().GetWidth(), (float) application.GetWindow().GetHeight());
 
 		// Rendering
 		ImGui::Render();
